Sherman's options now hide when other lang selected
