+++ js/doctorsCot/setup.js
@@ -118,42 +118,22 @@ export const letter = [];
     }
 });
 
-// generate IPA keyboard from letter tables
-export function initIpaKeys() {
-    const ipaConsons = document.getElementById('ipa-consons');
-    const ipaVowels = document.getElementById('ipa-vowels');
+// create HTML button elements laid out same as letter table
+export function genKeyboard(elem, table) {
+		for (let styl in table) {
+				const tableRow = table[styl].trim().split(/\s+/); // turn string to array, removing white space
+				const keyRow = document.createElement('div');
 
-    // move through table, adding button to IPA keyboard
-    function genKeyboard(elem, table) {
-        for (let styl in table) {
-            const tableRow = table[styl].trim().split(/\s+/); // turn string to array, removing white space
-            const keyRow = document.createElement('div');
+				tableRow.forEach(letter => {
+						const keyInput = document.createElement('input');
 
-            tableRow.forEach(letter => {
-                const keyInput = document.createElement('input');
+						keyInput.type = "button";
+						keyInput.value = letter;
+						keyInput.onclick = () =>
+								document.getElementById('text').value += letter;
 
-                keyInput.type = "button";
-                keyInput.value = letter;
-                keyInput.onclick = () =>
-                    document.getElementById('text').value += letter;
-
-                keyRow.appendChild(keyInput);
-            });
-            elem.appendChild(keyRow);
-        }
-    }
-
-    genKeyboard(ipaConsons, consonantTable);
-    genKeyboard(ipaVowels, vowelTable);
-
-    const ipaKeys = document.getElementById('ipa-keys');
-    const langSelect = document.getElementById('language');
-
-    // only show IPA keyboard when Doctor's Cot selected
-    langSelect.addEventListener('input', event => {
-        if (event.target.value == "cot")
-            ipaKeys.classList.remove('hidden');
-        else
-            ipaKeys.classList.add('hidden');
-    });
+						keyRow.appendChild(keyInput);
+				});
+				elem.appendChild(keyRow);
+		}
 }
