+++ main.css
@@ -3,8 +3,16 @@
     font-weight: bold;
 }
 
-.hidden {
+#lang-controls {
+    margin: .25em 0;
+    padding: 0;
+}
+#lang-controls > div {
     display: none;
+    padding: .5em;
+}
+#lang-controls > div.active {
+    display: block;
 }
